Moving mock API bootstrapping to environment.
